requirements.txt
anthropic>=0.49.0
streamlit>=1.41.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
plotly>=5.18.0
scipy>=1.11.0
networkx>=3.1
